Support for the (beta) Catalogs API
This PR adds support for the new Iterable Catalogs API, currently in beta.
> https://api.iterable.com/api/docs#catalogs_listCatalogs

This API is currently only available when using an API key that has been
granted beta access to catalogs.

Note: a lot of the functionality provided by this API is asynchronous,
so it's hard to guarantee that pre-conditions for the integration tests
have been properly set up without allowing time for the work to settle.

I've worked around this by adding a RECORD_DELAY environment variable
which will insert appropriate delays into the test suite to wait for the
async work to complete. This is far from ideal, but is only required when
recording new VCR cassettes.