Allow viewing an email by userId
Before viewing an email required specifying the email for the user of
the message. Using the user's `userId` wasn't an option, even though
Iterable's API supports this. This prevented userID-based projects
from querying the `viewInBrowser endpoint.

Similar to other `ApiResource`s, this changes `Iterable::Email#view`
to make email optional in favor of specifying the user's `userId` in
`attrs`, which sends with the request as query parameters.

Note: Some of the VCR cassettes were moved. The file contents is the
same, just a different location. This was to match file paths to the
block descriptions in the tests, which were reorganized so that they
would make sense with the added tests for using `userId`.